Fix the key input
There were two words to wait for a key, with the same name, but they did
not used the parameter. Now `?key-pause` is used.

It seems this problem had something to do with the recent errors that
happened at the start of the ambush or after finding the cave entry.
They have been fixed with the change.