Reorganize order of attributes and descriptions
This is a step in order to simplify the definition of entity
descriptions: `:description` will be removed, and the xt of a word will
be stored as an ordinary attribute, in `:attributes`. This will make
unnecessary to save the xt of de description before initializing the
entities data with the default values set by `:attributes`.