Simplify the definition of entity describers
`:description` has been removed. It worked fine, but it was only
syntactic sugar. Besides, it caused the xt of the describer need be
saved before executing the xt that initializes the attributes of the
entity (the `:attributes` definition).

An analogous simplification is planned for other xt that are not set in
the `:attributes` definition.