Replace dynamic string names with 2-cell strings
There was no need to use a dynamic string to store the name of the
entities, because the default string is already compiled in the
initialization word. Now the name field is a cell pair that stores the
compiled string. This little change makes some things simpler. For
example, now the only field that must be preserved during the
initialization of entities is the execution token of the initialization
word itself. This makes it easier to move the data structure to
Flibustre, because the application does not need to configure it.