Simplify the usage of the printing dynamic string
A temporary dynamic string (from Forth Foundation Library) is used by
some algorithms in order to make it easier to build the text that must
be printed at the end. The words used to manipulate the string were too
short and cryptical, and used Unicode characters. Now the words provided
by Forth Foundation Library are used instead. The function of one word
of the old interface is not provided by the library, so it has been
kept, modified  and renamed after it.